requirements.txt
python-telegram-bot
llama-cpp-python
apscheduler
duckduckgo_search
python-dotenv
sqlalchemy
sqlite3
